requirements.txt
streamlit>=1.35.0
pandas>=2.2.2
fpdf>=1.7.2
python-dotenv>=1.0.1
openai>=1.30.1
plotly>=5.21.0
matplotlib>=3.8.4
wordcloud>=1.9.3
textblob>=0.18.0
feedparser>=6.0.11
fpdf



